# Conflicts:
#	voyageone-service2/voyageone-service2-impl/src/main/java/com/voyageone/service/impl/cms/sx/SxProductService.java

+++ voyageone-task2/task2-cms/src/main/java/com/voyageone/task2/cms/service/monitor/ImageUploadService.java
@@ -11,9 +11,16 @@ import com.voyageone.common.configs.beans.FtpBean;
 import com.voyageone.common.util.DateTimeUtil;
 import com.voyageone.common.util.FtpUtil;
 import com.voyageone.common.util.JacksonUtil;
+import com.voyageone.common.util.StringUtils;
+import com.voyageone.service.impl.cms.BusinessLogService;
 import com.voyageone.service.impl.cms.product.ProductService;
+import com.voyageone.service.model.cms.CmsBtBusinessLogModel;
 import com.voyageone.service.model.cms.mongo.product.CmsBtProductModel;
 import com.voyageone.service.model.cms.mongo.product.CmsBtProductModel_Field;
+import com.voyageone.task2.base.Enums.TaskControlEnums;
+import com.voyageone.task2.base.dao.TaskDao;
+import com.voyageone.task2.base.modelbean.TaskControlBean;
+import com.voyageone.task2.base.util.TaskControlUtils;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.net.ftp.FTPClient;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,9 +44,18 @@ public class ImageUploadService extends AbstractFileMonitoService {
 
     private static final String S7FTP_CONFIG = "S7FTP_CONFIG";
 
+    private static final List<String> failedZipFileList = new ArrayList<>();
+
+    @Autowired
+    private BusinessLogService businessLogService;
+
+    @Autowired
+    private ProductService productService;
+
     @Override
-    protected void onCreate(String filePath) {
+    protected void onCreate(String filePath, String channelId) {
         LOG.info("监控到目录创建" + filePath);
+        onModify(filePath, channelId);
     }
 
     @Override
@@ -47,39 +63,73 @@ public class ImageUploadService extends AbstractFileMonitoService {
         LOG.info("监控到目录删除" + filePath);
     }
 
+    @Autowired
+    private TaskDao taskDao;
+
+
     @Override
     protected void onModify(String filePath, String channelId) {
+
+        List<TaskControlBean> taskControlList = taskDao.getTaskControlList(TASK_NAME);
+        TaskControlBean taskControlBean = TaskControlUtils.getVal1s(taskControlList, TaskControlEnums.Name.run_flg).get(0);
+
+        if (!"1".equals(taskControlBean.getCfg_val1())) {
+            LOG.warn(TASK_NAME + ":的run_flg设置为0,不执行该job");
+            return;
+        }
+
         LOG.info("监控到目录更新" + filePath);
         String tempDir = buildDirPath(filePath, "temp");
 //        String channelId = Properties.readValue(getClass().getSimpleName() + "_monitor_home_path_channelId");
         String modifyDirName = new File(filePath).getName();
         Arrays.asList(new File(filePath).listFiles((dir, name) -> name.endsWith(".zip"))).forEach(file -> {
-            try {
-                //解压缩到临时目录
-                deComparess(file, tempDir);
-                //依据临时目录内容上传ftp，之后更新Mongo数据
-                List<String> upFtpSuccessFiles = new ArrayList<String>();
-                ftpUpload(channelId, tempDir, buildDirPath(filePath, "error")).forEach((k, v) -> {
-                    if (v) upFtpSuccessFiles.add(k);
-                    else LOG.info(k + "Ftp上传失败");
-                });
-                //排序
-                Collections.sort(upFtpSuccessFiles);
-                upFtpSuccessFiles.forEach(k -> {
-                    CmsBtProductModel model = productService.getProductBySku(channelId, k.split("-")[2]);
-                    if (model != null) {
-                        mongoOperator(model, modifyDirName, k);
-                        approvedOperator(model);
+            if (!failedZipFileList.contains(file.getName())) {
+                int readImgCount = 0;
+                int failedImgCount = 0;
+                //异常信息
+                StringBuilder errorMsg = new StringBuilder();
+                try {
+                    //解压缩到临时目录
+                    deComparess(file, tempDir);
+                    //依据临时目录内容上传ftp，之后更新Mongo数据
+                    List<String> upFtpSuccessFiles = new ArrayList<>();
+                    Map<String, Boolean> ftpUploadResultMap = ftpUpload(channelId, tempDir, buildDirPath(filePath, "error"));
+                    readImgCount = ftpUploadResultMap.size();
+                    for (Map.Entry<String, Boolean> entry : ftpUploadResultMap.entrySet()) {
+                        if (entry.getValue()) {
+                            upFtpSuccessFiles.add(entry.getKey());
+                        } else {
+                            failedImgCount++;
+                            errorMsg.append(entry.getKey().split("-ftp-")[1] + "：图片上传到FTP服务器失败");
+                        }
                     }
-                });
-                //删除临时目录
-                FileUtils.deleteDirectory(new File(tempDir));
-                //移动文件到备份目录
-                FileUtils.moveFile(file, new File(buildDirPath(filePath, "backup") + file.getName() + "." + System.currentTimeMillis()));
-            } catch (IOException e) {
-                LOG.error("处理zip文件" + filePath + "发生Io异常：", e);
+                    //排序
+                    Collections.sort(upFtpSuccessFiles);
+                    Set<String> skuApprovedSet = new HashSet<>();
+                    upFtpSuccessFiles.forEach(k -> {
+                        CmsBtProductModel model = productService.getProductBySku(channelId, k.split("-")[2]);
+                        if (model != null) {
+                            mongoOperator(model, modifyDirName, k);
+                            if (skuApprovedSet.add(k.split("-")[2]))
+                                approvedOperator(model);
+                        } else {
+                            errorMsg.append(k.split("-ftp-")[1] + "：找不到指定商品");
+                        }
+                    });
+                    //删除临时目录
+                    FileUtils.deleteDirectory(new File(tempDir));
+                    //移动文件到备份目录
+                    FileUtils.moveFile(file, new File(buildDirPath(filePath, "backup") + file.getName() + "." + System.currentTimeMillis()));
+                } catch (IOException e) {
+                    failedZipFileList.add(file.getName());
+                    errorMsg.append("其他异常：" + e.getMessage());
+                    LOG.error("处理zip文件" + filePath + "发生Io异常：", e);
+                }
+                logForUpload(file.getName(), readImgCount, readImgCount - failedImgCount, failedImgCount, errorMsg.toString(), channelId);
             }
         });
+        taskControlBean.setEnd_time(DateTimeUtil.getNow());
+        taskDao.updateTaskControl(taskControlBean);
     }
 
     /**
@@ -181,16 +231,18 @@ public class ImageUploadService extends AbstractFileMonitoService {
      * @param dirFile  目录文件
      * @param fileList 文件list
      */
-    private static void filterImgFile(File dirFile, List<File> fileList) {
+    private void filterImgFile(File dirFile, List<File> fileList) {
         for (File file : dirFile.listFiles()) {
             if (file.isDirectory()) {
                 filterImgFile(file, fileList);
-            } else fileList.add(file);
+            } else if (file.getName().endsWith(".jpg") || file.getName().endsWith(".png") || file.getName().endsWith(".gif")
+                    || file.getName().endsWith(".jpeg")) {
+                fileList.add(file);
+            } else {
+                LOG.warn("不支持上传的的文件类型");
+            }
         }
     }
-
-    @Autowired
-    private ProductService productService;
 
     /**
      * 操作mongodb
@@ -202,14 +254,23 @@ public class ImageUploadService extends AbstractFileMonitoService {
 
             /* change fileds.images */
             List<Map<String, Object>> images = new CopyOnWriteArrayList<>();
-            if (!ObjectUtils.isEmpty(fields.get(modifyDirName)))
+            if (!ObjectUtils.isEmpty(fields.get(modifyDirName))) {
                 images = new CopyOnWriteArrayList<>(JacksonUtil.jsonToMapList(fields.get(modifyDirName).toString()));
+
+                // 只有一张图片,并且该图片的值为空的时候,删除其对应的图片信息
+                if (images.size() == 1)
+                    for(Map<String, Object> image : images) {
+                        StringUtils.isEmpty(String.valueOf(image.get(modifyDirName.replace("s", ""))));
+                        images.remove(image);
+                    }
+
+            }
             for (Map<String, Object> img : images) {
                 if (uploadFileName.equals(img.get(modifyDirName)))
                     return;
             }
             images.add(new HashMap<String, Object>() {{
-                put(modifyDirName.replace("s",""), uploadFileName.split("\\.")[0]);
+                put(modifyDirName.replace("s", ""), uploadFileName.split("\\.")[0]);
             }});
 
             /* 是images6 忽略大小写,重新排序 */
@@ -220,11 +281,11 @@ public class ImageUploadService extends AbstractFileMonitoService {
                 if ("1".equals(cmsChannelConfigBean.getConfigValue1()))
                     images.addAll(0, JacksonUtil.jsonToMapList(fields.get("images1").toString().replace("image1", "image6")));
                 if ("2".equals(cmsChannelConfigBean.getConfigValue1()))
-                    images.addAll(images.size(), JacksonUtil.jsonToMapList(fields.get("images1").toString().replace("image1", "image6")));
+                    images.addAll(JacksonUtil.jsonToMapList(fields.get("images1").toString().replace("image1", "image6")));
             }
 
-            Set<Map<String,Object>> sets=new HashSet<>();
-            for(Map<String,Object> map:images) if(!sets.add(map)) images.remove(map);
+            Set<Map<String, Object>> sets = new HashSet<>();
+            for (Map<String, Object> map : images) if (!sets.add(map)) images.remove(map);
 
             updateProductModel(model.getChannelId(), model.getProdId(), modifyDirName, images);
         } catch (IOException e) {
@@ -256,5 +317,28 @@ public class ImageUploadService extends AbstractFileMonitoService {
         queryMap.put("prodId", prodId);
 
         productService.updateProduct(channelId, queryMap, updateMap);
+    }
+
+    /**
+     * 上传日志
+     *
+     * @param zipName           zip文件名称
+     * @param readSkuCount      读取sku个数
+     * @param processedSkuCount 处理sku个数
+     * @param failedSkuCount    失败sku个数
+     * @param errorMsg          错误信息
+     */
+    private void logForUpload(String zipName, int readSkuCount, int processedSkuCount, int failedSkuCount, String errorMsg, String channelId) {
+        StringBuilder stringBuilder = new StringBuilder("ImageUploadService处理压缩文件" + zipName + "操作日志");
+        stringBuilder.append("\n本次读入图片个数：").append(readSkuCount);
+        stringBuilder.append("\n正常处理图片个数：").append(processedSkuCount);
+        stringBuilder.append("\n异常处理图片个数：").append(failedSkuCount);
+        stringBuilder.append("\n错误说明：").append(errorMsg);
+        CmsBtBusinessLogModel logModel = new CmsBtBusinessLogModel();
+        logModel.setChannelId(channelId);
+        logModel.setErrorMsg(stringBuilder.toString());
+        logModel.setErrorTypeId(2);
+        logModel.setCreater(TASK_NAME);
+        businessLogService.insertBusinessLog(logModel);
     }
 }
